fix(Cornerstone): Support mode-specific layout configs
This is in preparation for using a custom Layout2D in conjunction with
Cornerstone.
