+++ Sources/config/Medical/layout.js
@@ -1,0 +1,3 @@
+import Generic from '../Generic';
+
+export default Generic.Layout;
